requirements.txt
Flask
requests
pprint
prettytable
redis
python-daemon
